+++ Firestore/core/src/local/memory_lru_reference_delegate.cc
@@ -116,7 +116,7 @@ size_t MemoryLruReferenceDelegate::GetSequenceNumberCount() {
 int MemoryLruReferenceDelegate::RemoveTargets(
     model::ListenSequenceNumber sequence_number,
     const LiveQueryMap& live_queries) {
-  return persistence_->target_cache()->RemoveTargets(sequence_number,
+  return (int)persistence_->target_cache()->RemoveTargets(sequence_number,
                                                      live_queries);
 }
